+++ source/marketparticipant/Energinet.DataHub.MarketParticipant.EntryPoint.WebApi/Controllers/ActorContactController.cs
@@ -13,8 +13,11 @@
 // limitations under the License.
 
 using System;
+using System.Collections.Generic;
 using System.Threading.Tasks;
+using Energinet.DataHub.Core.App.Common.Abstractions.Users;
 using Energinet.DataHub.MarketParticipant.Application.Commands.Contact;
+using Energinet.DataHub.MarketParticipant.Application.Security;
 using Energinet.DataHub.MarketParticipant.Domain.Model.Permissions;
 using Energinet.DataHub.MarketParticipant.EntryPoint.WebApi.Security;
 using MediatR;
@@ -27,15 +30,17 @@ namespace Energinet.DataHub.MarketParticipant.EntryPoint.WebApi.Controllers
     public sealed class ActorContactController : ControllerBase
     {
         private readonly IMediator _mediator;
+        private readonly IUserContext<FrontendUser> _userContext;
 
-        public ActorContactController(IMediator mediator)
+        public ActorContactController(IMediator mediator, IUserContext<FrontendUser> userContext)
         {
             _mediator = mediator;
+            _userContext = userContext;
         }
 
         [HttpGet("{actorId:guid}/contact")]
         [AuthorizeUser(PermissionId.ActorsManage)]
-        public async Task<IActionResult> ListAllAsync(Guid actorId)
+        public async Task<ActionResult<IEnumerable<ActorContactDto>>> ListAllAsync(Guid actorId)
         {
             var getOrganizationsCommand = new GetActorContactsCommand(actorId);
 
@@ -48,8 +53,11 @@ namespace Energinet.DataHub.MarketParticipant.EntryPoint.WebApi.Controllers
 
         [HttpPost("{actorId:guid}/contact")]
         [AuthorizeUser(PermissionId.ActorsManage)]
-        public async Task<IActionResult> CreateContactAsync(Guid actorId, CreateActorContactDto contactDto)
+        public async Task<ActionResult<string>> CreateContactAsync(Guid actorId, CreateActorContactDto contactDto)
         {
+            if (!_userContext.CurrentUser.IsFasOrAssignedToActor(actorId))
+                return Unauthorized();
+
             var createContactCommand = new CreateActorContactCommand(actorId, contactDto);
 
             var response = await _mediator
@@ -61,8 +69,11 @@ namespace Energinet.DataHub.MarketParticipant.EntryPoint.WebApi.Controllers
 
         [HttpDelete("{actorId:guid}/contact/{contactId:guid}")]
         [AuthorizeUser(PermissionId.ActorsManage)]
-        public async Task<IActionResult> DeleteContactAsync(Guid actorId, Guid contactId)
+        public async Task<ActionResult> DeleteContactAsync(Guid actorId, Guid contactId)
         {
+            if (!_userContext.CurrentUser.IsFasOrAssignedToActor(actorId))
+                return Unauthorized();
+
             var deleteContactCommand = new DeleteActorContactCommand(actorId, contactId);
 
             await _mediator