+++ source/marketparticipant/Energinet.DataHub.MarketParticipant.EntryPoint.WebApi/Startup.cs
@@ -160,6 +160,8 @@ namespace Energinet.DataHub.MarketParticipant.EntryPoint.WebApi
                 };
 
                 c.AddSecurityDefinition("Bearer", securitySchema);
+                c.SupportNonNullableReferenceTypes();
+                c.UseAllOfToExtendReferenceSchemas();
 
                 var securityRequirement = new OpenApiSecurityRequirement { { securitySchema, new[] { "Bearer" } } };