+++ source/api-client/Energinet.DataHub.MarketParticipant.Client/documents/release-notes/release-notes.md
@@ -1,5 +1,13 @@
 # Energinet.DataHub.MarketParticipant.Client Release notes
 
+## Version 2.33.0
+
+- Remove assigned actors from user overview.
+
+## Version 2.32.6
+
+- API for resetting users 2FA.
+
 ## Version 2.32.5
 
 - Added domain to OrganizationChangeDto